+++ lua/obsidian/commands/init-legacy.lua
@@ -138,7 +138,7 @@ M.register("ObsidianTomorrow", { opts = { nargs = 0, desc = "Open the daily note
 
 M.register("ObsidianDailies", { opts = { nargs = "*", desc = "Open a picker with daily notes" } })
 
-M.register("ObsidianNew", { opts = { nargs = "?", complete = "file", desc = "Create a new note" } })
+M.register("ObsidianNew", { opts = { nargs = "?", desc = "Create a new note" } })
 
 M.register(
   "ObsidianOpen",
@@ -172,15 +172,9 @@ M.register("ObsidianToggleCheckbox", { opts = { nargs = 0, desc = "Toggle checkb
 
 M.register("ObsidianWorkspace", { opts = { nargs = "?", desc = "Check or switch workspace" } })
 
-M.register(
-  "ObsidianRename",
-  { opts = { nargs = "?", complete = "file", desc = "Rename note and update all references to it" } }
-)
+M.register("ObsidianRename", { opts = { nargs = "?", desc = "Rename note and update all references to it" } })
 
-M.register(
-  "ObsidianPasteImg",
-  { opts = { nargs = "?", complete = "file", desc = "Paste an image from the clipboard" } }
-)
+M.register("ObsidianPasteImg", { opts = { nargs = "?", desc = "Paste an image from the clipboard" } })
 
 M.register(
   "ObsidianExtractNote",