+++ lua/obsidian/commands/init.lua
@@ -209,7 +209,7 @@ M.register("tomorrow", { nargs = 0 })
 
 M.register("dailies", { nargs = "*" })
 
-M.register("new", { nargs = "?", complete = "file" })
+M.register("new", { nargs = "?" })
 
 M.register("open", { nargs = "?", complete = M.note_complete })
 
@@ -237,9 +237,9 @@ M.register("toggle_checkbox", { nargs = 0, range = true })
 
 M.register("workspace", { nargs = "?" })
 
-M.register("rename", { nargs = "?", complete = "file" })
+M.register("rename", { nargs = "?" })
 
-M.register("paste_img", { nargs = "?", complete = "file" })
+M.register("paste_img", { nargs = "?" })
 
 M.register("extract_note", { nargs = "?", range = true })