+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `<C-x>` not working after command merging
 - Fixed `blink.cmp` new notes source completion not working properly
 - Fixed `blink.cmp` off by one insert start position
+- Fixed `ObsidianNew`, `ObsidianRename`, and `ObsidianPasteImg` not to evaluate backticked substrings in its arguments, disabling `-complete=file`
 
 ## [v3.10.0](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.10.0) - 2025-04-12